ExamOnlineAPI/requirements.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
celery==5.4.0
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
coverage==7.6.9
cryptography==44.0.0
Django==5.1.4
django-environ==0.11.2
djangorestframework==3.15.2
iniconfig==2.0.0
kombu==5.4.2
mysqlclient==2.2.6
packaging==24.2
pdfminer.six==20231228
pdfplumber==0.11.4
pillow==11.0.0
pluggy==1.5.0
prompt_toolkit==3.0.48
psycopg2-binary==2.9.10
pycparser==2.22
pypdfium2==4.30.1
pytest==8.3.4
pytest-django==4.9.0
pytest-mock==3.14.0
python-dateutil==2.9.0.post0
six==1.17.0
sqlparse==0.5.3
tzdata==2024.2
vine==5.1.0
wcwidth==0.2.13
